requirements.txt
# ===============================================
# DEPENDENCIAS PROYECTO WPC PYTHON
# ===============================================

# Base de datos SQL Server
pyodbc>=4.0.39
sqlalchemy>=2.0.23
alembic>=1.12.1

# Comunicación serie RS485
pyserial>=3.5

# Interfaz gráfica (elegir una)
PyQt6>=6.6.0
# PySide6>=6.6.0  # Alternativa a PyQt6

# Integración cámaras Hikvision
requests>=2.31.0
opencv-python>=4.8.1.78

# Procesamiento de imágenes
Pillow>=10.1.0
numpy>=1.24.0

# Configuración y utilidades
python-dotenv>=1.0.0
loguru>=0.7.2

# Comunicación asíncrona
asyncio>=3.4.3

# Fecha y hora
python-dateutil>=2.8.2

# Validación de datos
pydantic>=2.5.0

# Testing y desarrollo
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Formateo y linting
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0

# Documentación
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# Monitoreo y métricas
psutil>=5.9.0

# Criptografía (para contraseñas de cámaras)
cryptography>=41.0.0

# Interfaz web (opcional para panel de administración)
fastapi>=0.104.0
uvicorn>=0.24.0

# Exportación de datos
openpyxl>=3.1.0
pandas>=2.1.0